refactor: extract cycle detection into analyzer/cycles.rs (#190)
Moves DetectedCycle (pub(super)), detect_sibling_cycles, find_sccs
(Tarjan's iterative SCC), and find_cycle_in_scc (DFS for directed cycle
in an SCC) out of analyzer/mod.rs. mod.rs now imports
detect_sibling_cycles and uses it in audit()'s BFS sweep.
